backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
python-multipart==0.0.6
librosa==0.10.1
scikit-learn==1.3.2
xgboost==2.0.2
pydantic==2.5.2
numpy==1.26.2
pandas==2.1.3
fpdf2==2.7.7
shap==0.44.0
matplotlib==3.8.2
